Code Cleanup and Some refactoring
Added for climbing options. When the robot is tilted enough, Ty can
take over with his joysticks and control each side of the robot. Grant
can also control both with his joystick. But that is only if the
SMartDashboard option is set to Joystick. If set to button, then the
climber just keeps going until the button it clicked again