More filtering of shapes
And test photos too. Each part of the filtering is a different color
when drawn on the screen. For photos, it stops after each one and waits
for a keypress. Video and camera feeds it doesn’t do that. Need to
verify timing is still ok. Also, the checks could be ordered better and
cleaned up a bit.